Remove logback as a main dependency
Popular Java logging technologies like SLF4J/Logback, and Log4J2, split the process of logging into 2 components: a logging frontend, and a logging backend. The logging frontend is the API that libraries and applications use to notify the logging system that they have something to report. The logging backend is used only by applications, not libraries, to determine how and where to log the stream of events that are coming from application code, as well as various libraries.

SLF4J is a logging frontend. Logback is a logging backend.

The popular logging technologies have adapters to work with one another. For example, Log4J2 can be used as a backend for SLF4J, and Logback can be used as a backend for JUL.

This library has chosen to code against the SLF4J API to report events that should be logged. However, this library should not dictate to consuming applications the logging backend they should use to process the log events, which it does by declaring a dependency on Logback. Accordingly, this commit removes the main dependency on Logback.